Refactor getUser so it can be used in user service
